[charts] Extract MapImagePlot reprojection into a documented helper
Move the per-pixel inverse-resampling logic out of the effect into reprojectEquirectangularImage, with a thorough explanation of the inverse-mapping approach, the round-trip footprint test, and the failure modes.
