+++ packages/x-charts-premium/src/Map/MapImagePlot.tsx
@@ -4,6 +4,7 @@ import PropTypes from 'prop-types';
 import { useDrawingArea } from '@mui/x-charts/hooks';
 import type { GeoProjection } from '@mui/x-charts-vendor/d3-geo';
 import { useGeoPath } from '../hooks/useGeoPath';
+import { reprojectEquirectangularImage } from './reprojectEquirectangularImage';
 
 const FULL_GLOBE: [[number, number], [number, number]] = [
   [-180, -90],
@@ -55,7 +56,6 @@ function MapImagePlot(props: MapImagePlotProps) {
       setDataUrl(null);
       return undefined;
     }
-    const { invert } = projection;
 
     let cancelled = false;
     const image = new Image();
@@ -64,72 +64,17 @@ function MapImagePlot(props: MapImagePlotProps) {
       if (cancelled) {
         return;
       }
-      const outWidth = Math.max(1, Math.round(width));
-      const outHeight = Math.max(1, Math.round(height));
-
-      const source = document.createElement('canvas');
-      source.width = image.naturalWidth;
-      source.height = image.naturalHeight;
-      const sourceCtx = source.getContext('2d');
-      const output = document.createElement('canvas');
-      output.width = outWidth;
-      output.height = outHeight;
-      const outputCtx = output.getContext('2d');
-      if (!sourceCtx || !outputCtx) {
-        return;
-      }
-      sourceCtx.drawImage(image, 0, 0);
-
-      let sourcePixels: ImageData;
-      try {
-        sourcePixels = sourceCtx.getImageData(0, 0, source.width, source.height);
-      } catch {
-        // Cross-origin image without CORS taints the canvas: bail out.
-        return;
-      }
-
-      const sourceWidth = source.width;
-      const sourceHeight = source.height;
-      const target = outputCtx.createImageData(outWidth, outHeight);
-
-      for (let py = 0; py < outHeight; py += 1) {
-        for (let px = 0; px < outWidth; px += 1) {
-          const deviceX = left + px;
-          const deviceY = top + py;
-          const coordinates = invert([deviceX, deviceY]);
-          if (!coordinates) {
-            continue;
-          }
-          const [lon, lat] = coordinates;
-          if (lon < west || lon > east || lat < south || lat > north) {
-            continue;
-          }
-          // Skip pixels hidden by the projection: if the coordinate does not
-          // project back to this pixel, it lies outside the visible footprint.
-          const reprojected = projection([lon, lat]);
-          if (
-            !reprojected ||
-            Math.abs(reprojected[0] - deviceX) > 0.5 ||
-            Math.abs(reprojected[1] - deviceY) > 0.5
-          ) {
-            continue;
-          }
-          let sx = Math.floor(((lon - west) / (east - west)) * sourceWidth);
-          let sy = Math.floor(((north - lat) / (north - south)) * sourceHeight);
-          sx = Math.min(Math.max(sx, 0), sourceWidth - 1);
-          sy = Math.min(Math.max(sy, 0), sourceHeight - 1);
-
-          const sourceIndex = (sy * sourceWidth + sx) * 4;
-          const targetIndex = (py * outWidth + px) * 4;
-          target.data[targetIndex] = sourcePixels.data[sourceIndex];
-          target.data[targetIndex + 1] = sourcePixels.data[sourceIndex + 1];
-          target.data[targetIndex + 2] = sourcePixels.data[sourceIndex + 2];
-          target.data[targetIndex + 3] = sourcePixels.data[sourceIndex + 3];
-        }
-      }
-
-      outputCtx.putImageData(target, 0, 0);
-      setDataUrl(output.toDataURL());
+      setDataUrl(
+        reprojectEquirectangularImage({
+          image,
+          projection,
+          area: { left, top, width, height },
+          imageBounds: [
+            [west, south],
+            [east, north],
+          ],
+        }),
+      );
     };
     image.src = href;
